fix: avoid gateway health import cycle
